bug fix class Review IntegerField

+++ base_info_app/api/serializers.py
@@ -3,6 +3,6 @@ from rest_framework import serializers
 
 class BaseInfoSerializer(serializers.Serializer):
     review_count = serializers.IntegerField()
-    average_rating = serializers.DecimalField(max_digits=2, decimal_places=1)
+    average_rating = serializers.FloatField()
     business_profile_count = serializers.IntegerField()
     offer_count = serializers.IntegerField()
